Update Pro-Guard description
Do you mean the Pro-Guard rules are required for 0.9.2 and ABOVE? I used 1.0.1 without Pro-Guard rules but then nothing worked in the release build. Using 1.0.1 with the Pro-Guard rules everything worked fine.

So I think the statement should be changed from "The following configuration is only needed for version 0.9.2 and below" to "The following configuration is only needed for version 0.9.2 and above".